requirements.txt
langchain-community
langchain[google-genai]
langchain-google-genai
langchain-qdrant
langgraph
langchain-text-splitters
qdrant-client
python-dotenv
